NLRFIM-41 README update for running all tests + rest refactor
