+++ integrationtest/rest/src/test/java/signature/RestSignatureScenarioDotnetTest.java
@@ -11,19 +11,7 @@ import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringRunner.class)
 @ContextConfiguration(locations = {"file:../../examples/oio-idws-rest/rest-client/src/main/resources/cxf.xml"})
-public class RestSignatureScenarioTest {
-
-    @Test
-    public void testSignatureScenario_toJavaWsp() {
-        Application.setRequestUrl("https://localhost:8443/api/hello?name=John");
-        Application.setTokenUrl("https://localhost:8443/auth");
-        Application.setAudience("https://wsp.itcrew.dk");
-        SpringApplication.run(Application.class);
-        String restResponse = Application.getRestResponse();
-
-        assertTrue(restResponse.contains("200"));
-        assertTrue(restResponse.contains("Hello John"));
-    }
+public class RestSignatureScenarioDotnetTest {
 
     //For local testing run Digst.OioIdws.Rest.Examples.ServerAndASCombined
     @Test
@@ -40,4 +28,5 @@ public class RestSignatureScenarioTest {
         assertTrue(restResponse.contains("AuthenticationType: OioIdws"));
         assertTrue(restResponse.contains("Claims"));
     }
+
 }
